+++ generators/app/index.js
@@ -1,18 +1,9 @@
 const yeoman = require('yeoman-generator');
-// Const chalk = require('chalk');
-// const ejs = require('ejs');
-// const glob = require('glob');
-// const mkdirp = require('mkdirp');
-// const findParentDir = require('find-parent-dir');
-// const fs = require('fs');
-// const spawn = require('cross-spawn');
-// const generatorPackageJson = require('../../package.json');
-// const { FILE_DELIM_OPEN, FILE_DELIM_CLOSE } = require('../util/ejs-util');
 
 module.exports = yeoman.generators.Base.extend({
   prompting: function() {
     const done = this.async();
-    this.prompt(
+    const prompts = [
       {
         type: 'input',
         name: 'name',
@@ -20,6 +11,20 @@ module.exports = yeoman.generators.Base.extend({
         // Defaults to the project's folder name if the input is skipped
         default: this.appname
       },
+      {
+        type: 'input',
+        name: 'description',
+        message: 'Write a brief description for your component?'
+      },
+      {
+        type: 'input',
+        name: 'author',
+        message: 'What is your name?'
+      }
+    ];
+
+    this.prompt(
+      prompts,
       function(answers) {
         this.props = answers;
         this.log(answers.name);
@@ -35,8 +40,30 @@ module.exports = yeoman.generators.Base.extend({
         this.templatePath('package.json'),
         this.destinationPath('package.json'),
         {
+          name: this.props.name,
+          description: this.props.description,
+          author: this.props.author
+        }
+      );
+      this.fs.copyTpl(this.templatePath('README.md'), this.destinationPath('README.md'), {
+        name: this.props.name
+      });
+      this.fs.copyTpl(
+        this.templatePath('CHANGELOG.md'),
+        this.destinationPath('CHANGELOG.md'),
+        {
           name: this.props.name
         }
+      );
+      this.fs.copyTpl(
+        this.templatePath('webpack.config.js'),
+        this.destinationPath('webpack.config.js'),
+        {}
+      );
+      this.fs.copyTpl(
+        this.templatePath('PULL_REQUEST_TEMPLATE.md'),
+        this.destinationPath('PULL_REQUEST_TEMPLATE.md'),
+        {}
       );
 
       // Copy application files